testserver: move from internal to pkg
