+++ controllers/gitrepository_controller_test.go
@@ -39,7 +39,7 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 
 	sourcev1 "github.com/fluxcd/source-controller/api/v1alpha1"
-	"github.com/fluxcd/source-controller/internal/testserver"
+	"github.com/fluxcd/source-controller/pkg/testserver"
 )
 
 var _ = Describe("GitRepositoryReconciler", func() {
