+++ controllers/helmchart_controller_test.go
@@ -31,7 +31,7 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 
 	sourcev1 "github.com/fluxcd/source-controller/api/v1alpha1"
-	"github.com/fluxcd/source-controller/internal/testserver"
+	"github.com/fluxcd/source-controller/pkg/testserver"
 )
 
 var _ = Describe("HelmChartReconciler", func() {
